pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["code-begemot <code.begemode@gmail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-bootstrap5 = "^24.2"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
django-filter = "^24.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
rollbar = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
